# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Clover_TH API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/auth/discord/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512
            },
            "in": "query",
            "name": "code",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "query",
            "name": "state",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 0,
              "maxLength": 128
            },
            "in": "query",
            "name": "error",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/bot/members/{discordId}": {
      "put": {
        "tags": [
          "bot"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ign": {
                    "type": "string",
                    "maxLength": 256
                  },
                  "job": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "jobId": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991
                  },
                  "nickname": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 256
                  }
                },
                "required": [
                  "ign"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{5,25}$"
            },
            "in": "path",
            "name": "discordId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memberId": {
                      "type": "string"
                    },
                    "discordId": {
                      "type": "string"
                    },
                    "ign": {
                      "type": "string"
                    },
                    "nickname": {
                      "nullable": true,
                      "type": "string"
                    },
                    "job": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "label": {
                          "type": "string"
                        },
                        "color": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "label",
                        "color"
                      ],
                      "additionalProperties": false
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "isIncomplete": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "memberId",
                    "discordId",
                    "ign",
                    "nickname",
                    "job",
                    "isActive",
                    "isIncomplete"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memberId": {
                      "type": "string"
                    },
                    "discordId": {
                      "type": "string"
                    },
                    "ign": {
                      "type": "string"
                    },
                    "nickname": {
                      "nullable": true,
                      "type": "string"
                    },
                    "job": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "label": {
                          "type": "string"
                        },
                        "color": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "label",
                        "color"
                      ],
                      "additionalProperties": false
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "isIncomplete": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "memberId",
                    "discordId",
                    "ign",
                    "nickname",
                    "job",
                    "isActive",
                    "isIncomplete"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bot/members/{discordId}/deactivate": {
      "post": {
        "tags": [
          "bot"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{5,25}$"
            },
            "in": "path",
            "name": "discordId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memberId": {
                      "type": "string"
                    },
                    "isActive": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  },
                  "required": [
                    "memberId",
                    "isActive"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}